+++ spec/examples/factory_example_spec.rb
@@ -5,9 +5,7 @@ require './examples/factory_example'
 
 rast FactoryExample do
   spec '#phone_plan_name' do
-    execute do |service_type|
-      subject.instance_variable_set(:@phone, build(service_type.to_sym))
-      subject.phone_plan_name
-    end
+    prepare { |service_type| subject.instance_variable_set(:@phone, build(service_type.to_sym)) }
+    execute { subject.phone_plan_name }
   end
 end