+++ spec/spec_helper.rb
@@ -103,6 +103,8 @@ RSpec.configure do |config|
   #   # as the one that triggered the failure.
   #   Kernel.srand config.seed
 
+  config.include FactoryGirl::Syntax::Methods
+
   config.before(:suite) do
     FactoryGirl.find_definitions
   end